Nit in Cells:
Towards a base class - that code was blacklisted, not least of all for
'size'. get_all_signatures might be cached on load?

Note iter_records returns stuff in correct order now. However this list
of lists is essentially a list of Mob*s - aka groups of records. This
needs lots of work, inheritance here is a mess.

The comment in MobBase.dump doesn't seem to apply anymore.