WIP setCell - fix for loading MobCells:
Seems we were using the header from the last iteration while the header
should be a groupType = 6 one as seen in setCell - and a dummy one if no
children -> confirmed see discussion in:

https://github.com/wrye-bash/wrye-bash/commit/dce017a5e8a37975320315c4ee905a235da9953e

from where:

>  And actually yeah, that might be exactly what's going on here: a header from
>  an interior or exterior cell being used for one with the opposite type,
>  leading to BSBs for interiors (type is (int, int)) being compared to BSBs for
>  exteriors (type is ((int, int), (int, int)).

So this probably closes #613

Moral: Beware of closures mein Sohn - we both looked at this code and
couldn't figure it out - keep closures a couple lines. Plus if it takes
20 mins to grok a closure, there is a bug and only way to grok is to
rewrite.

Fix for WorldBlocks headers

- load_rec_group are very similar now - what's the deal with isFallout??

Nit in _load_rec_group

We need a method in the header maybe