Add support for Fallout 3 on EGS
Would you like a rerelease?

I do not like it, Todd-You-Tease.
I do not like a rerelease.

Would you like it here or there?

I would not like it here or there.
I would not like it anywhere.
I do not like a rerelease.
I do not like it, Todd-You-Tease.

Under #648