Fix INI scrolling logic
Was completely wrong: for a start, it was using the index from the tweak
list inside the main INI list. But more the point, the correct logic is
actually super easy: get the current scroll position in the main INI
list, scroll up *minus* that many indices (this puts us at the top of
the list), then scroll down to the index we want. We can then, as an
optimization, do both of those in one scroll by just adding the negative
offset and the final index together.

The rest is me cleaning up Freeze/Thaw calls - they must always occur as
a pair, even if an exception got raised inbetween (which our de-wx'd
_ACFrozen handler as used by pause_drawing() already does correctly).