[!]MobDial(TopComplexGrup):
Whereupon the first setRecord for a complex group emerges - I wish it
was as "simple"  for the CELL groups also.

Note peculiar getSize - what I suspect is that when we created an INFO
collection for the BP top DIAL record we create the header also with
stamp=0 - the dialogue, being a MobObjects (...), carried the header of
the INFO records and would dump this using its stamp - hence the getSize
override. Note:

    def _mob_header(self, *args): # FIXME was TESIV only - self.extra??
        return self._grup_header_type(self.getSize(), self.label, self.stamp)

The self.stamp is the MobObjects stamp. Now all this needs verification
that can wait till actually merge dialogue records (note also the big
fat FIXME - we should revisit what we write to these headers on newer
games).

Error messages/str/repr is WIP

For _process_rec see https://stackoverflow.com/a/26538329/281545

[!]:

@@ -254,5 +260,3 @@ def dump(self,out):
         if not self.changed:
-            out.write(self._grup_header_type(self.size, self.label,
-                                    ##: self.header.pack_head() ?
-                                    self.stamp).pack_head())
+            out.write(self.header.pack_head())
             out.write(self.data)

_ComplexRec/required refactoring and WIP thin _ChildrenGrup:

TODO: I made MobObjects._load_rec_group a bit slower - but I don't think
overriding just to avoid the if check is worth it. Essentially the if
will be always false, given loadFactory will add all the cell record
classes if one is present - so this needs some more thought.

Move some of the machinery of MobDial to _ComplexRec

Edit comments for now they are still useful to track where that's from
(done later in the Some renaming commit)

WIP - note subtle differences with Cell (duh) - I changed to diff easier

Introduced rewind for readability

Infernio:

Fix keepRecords for WRLD records EEE

This might not be the right way to fix it, but the problem was that it
was using iter_records(), which also returned the WRLD we're trying to
check the children of. Not sure why it worked with CELL, might've just
been an accident of us not copying in any CELLs that we don't also
keep().

@@ -587,6 +587,13 @@ def _load_rec_group(self, ins, endPos, master_rec=None):
     def keepRecords(self, p_keep_ids):
-        if any(self.iter_records()):
-            p_keep_ids.add(self.master_record.group_key())
-        elif self.master_record.group_key() not in p_keep_ids:
+        if mr_rec := self.master_record:
+            # Get rid of the master record and look for children
             self.master_record = None
+            any_children_kept = any(self.iter_records())
+            if any_children_kept or mr_rec.group_key() in p_keep_ids:
+                # Either we want to keep the master record or at least one of
+                # its children, so restore the master record and keep it if it
+                # isn't already kept. Otherwise, just don't undo the master
+                # record deletion from before
+                p_keep_ids.add(mr_rec.group_key())
+                self.master_record = mr_rec

Utumno: could be changed to:

@@ -588,12 +588,10 @@ def keepRecords(self, p_keep_ids):
         if mr_rec := self.master_record:
-            # Get rid of the master record and look for children
-            self.master_record = None
-            any_children_kept = any(self.iter_records())
-            if any_children_kept or mr_rec.group_key() in p_keep_ids:
+            records = self.iter_records()
+            next(records) # Get rid of the master record and look for children
+            if any(records) or mr_rec.group_key() in p_keep_ids:
                 # Either we want to keep the master record or at least one of
-                # its children, so restore the master record and keep it if it
-                # isn't already kept. Otherwise, just don't undo the master
-                # record deletion from before
+                # its children, so keep the master record
                 p_keep_ids.add(mr_rec.group_key())
-                self.master_record = mr_rec
+            else:
+                self.master_record = None

cause this king of thing might need a finally block - we need to make
sure iter_records returns an iterator always (returns values() once and
this would blow with next()).