Interlude - RecordType:
Our predecessors meant something with that recType thing - however we
needed to wait a couple decades till py3 gave us the right tools. Does
not do much yet, but will. Anyway repetitive code in the class
definitions is a sure sign of a metaclass.

Idea from https://stackoverflow.com/a/20086971/281545 and

https://pyvideo.org/pycon-us-2013/python-3-metaprogramming.html

:P