Move validate_name to ListInfo:
Was needed for Inf's refactoring of the Ini tab and was high time anyway.
It needed info_dirs - moved that too - note ListInfo warns on unresolved
abs_path - that is True, ListInfo is not a file necessarily (default
tweaks) - still we can fix that warning last - note current API is
already expressive enough to even cater for the horribleness of
info_dir (it seems to suggest that we can stash an info in whatever dir
while that is not quite true - store_dir is really supported).

Mopy/bash/basher/installer_links.py: Path.open will make the dirs if
needed

Under #336.