Rework date and time input
Create a proper popup that lets users pick via a higher-level interface
while also maintaining the old 'manual' entry method, but adding
immediate validation so that users know when they've entered an invalid
date.

Could have applied this to the 'Modified' text field on the Mods tab
too, but that's a text field and I think a text field opening a whole
popup when clicked is really weird behavior. I know polemos' fork does
that, but I found it weird and jarring when I first saw it - and I'm
trying to make WB *less* weird and jarring :P
So let's keep the 'Modified' text field as is - it's probably very
rarely used anyways and can just remain an optional, potentially
quicker way to set the timestamp manually.

Also dropped unformat_date, that workaround isn't necessary in py3.