WIP LoadFactory:
Follow up for 645015dfb680c4aecf1408919d5b13e0dd2d7dfe:

- I removed recTypes (maybe makes sense to differentiate between
MreRecord and MelRecord ones?)
- this permitted me to slot LF - we need to be as fast as possible
accessing the attributes of LF, especially in the context of record
groups
- reinstated valid_header_sigs - we needed some more validation,
especially in _TopGrupDict where the validation happens in TopGrupHeader
which will change further down this branch
- validation and responsibilities are yet a bit fluid and may happen
twice in some of the paths - ModFile load vs _TopGrupDict __missing__
for instance

But the most important thing here is moving the handling of top group
loading in _import_records. This is what really makes LF what it
should be - game agnostic.