Add a button for interactive mtime changes
Middle ground between doing nothing about the 'Modified' text field and
doing the "magic" behavior of polemos' fork. This adds a small button
with a calendar icon at the end of the 'Modified' field that launches
the date and time dialog when pressed.

Also fixed a bug that was causing the error message for invalid manually
entered dates to fail.

Under #479 for parity with polemos' Wrye Mash fork.