Oblivion versions handling move to game:
Note that the versions strings were all over the place - and reOblivion
was also checking for the 1.5.0.8 version that was not included in the
dicts. This centralizes version handling in the game package where it
belongs, cleans up ModInfos a bit, gets us rid of reOblivion and is
still WIP as I would like to avoid all those sting operations for
readability (I would like to use the modding esm names rather than the
version strings). Note that in patch_files.py would be more correct to
use the voAvailable but then let's say that those modding esm names are
reserved. Anyway that's good enough for now.

Under #649

Correctly refresh the menu state

Infernio:

Move Oblivion.esm link init logic to mods_links

I honestly want less game-specific logic and such in links_init - IMO
that file should just define the layout, i.e. the order and nesting of
menus and links, as well as separators. That also lets us avoid
duplicate logic for making links game-specific.

Found an easy way to add the separator back in, so here it is.