Rework Installers icons handling
Started because without it, the ImageList will contain images with
different sizes on HiDPI, which causes giant warning popups on startup
('Failed to add image to ImageList' or something like that) and crashes
when one opens and closes the Settings menu.

In the process this turned out to let us get rid of installercons and
with it some of the weird interdependence in basher, which is very
welcome.

There is more work to be done - the ImageWrapper calls I threw into
dialogs.py should instead really access balt.images, so that they're
amenable to customizable checkboxes in the future, same with ColorChecks
and InstallerColorChecks - but all that is pt2 stuff.