Fix Renumber Saves not working and improve name
Re-number Save(s) is a grammar mistake (renumber, not re-number) and due
to the usage of '(s)' hard/impossible to translate into some languages.
Let's just call it Renumber Saves and be done with it.

Also fixes several incorrect usages of the rename APIs that caused it to
add duplicate extensions and to only work for the first selected save.