bain: Add redirect for top-level CSVs
Requested on Discord. Many CSVs for use with the BP in Oblivion are
distributed with the CSVs simply at the top level rather than inside the
Bash Patches folder, where they have to be for WB to recognize them. The
intention is probably to have people manually install them or the
authors in question are unaware that BAIN can install CSVs as long as
they are placed in a Bash Patches folder. This redirect makes such
packages work directly in BAIN.

Also slightly cleaned up affected code in the files and readme.