Show warning when installing unconfigured packages
This is for new users. We have a great first time user experience now
w.r.t. installing packages for most package types:

 - Simple packages: right click > Install... just works:tm:
 - BAIN wizards: right click > Install... just works:tm:
 - FOMODs: right click > Install... just works:tm:

The exception is Complex packages that don't have a BAIN wizard or
FOMOD, where right click > Install... will silently install nothing at
all. Which is because you have to select sub-packages first, of course.
So show a popup telling users exactly that when we get into this
situation.

Also renamed all usages of 'subpackage' to 'sub-package' as a minor
aesthetic thing.