Make the Filter tag's behavior more sensible
Just about everyone seems to be confused as to what this tag actually
does - LOOT, our ancestors who wrote this code, the docs, us... See the
issue linked below for more information.

So the final consensus is: Filter is used if you have a plugin with
missing masters and want to include data from it into the BP, whether by
importing (using preservers and/or mergers) or by merging (using Merge
Patches). If you don't tag your plugin with it and it has missing
masters, it will be skipped instead.

Closes #654