Remove the Eye Checker
Unless someone can provide me with an actual case where this thing helps
rather than blows everything up, we can talk about writing a replacement
*from scratch*.

Because as it stands, all this does is break things. As far as I
understand the issue ('googly eyes'), it happens as follows (based on
information from https://www.nexusmods.com/oblivion/mods/46254):
 - Mod A changes an eye mesh
 - Mod B changes an eye texture
 - Due to the different UV mapping of Mod A's eye mesh, Mod B's eye
   texture is no longer compatible with the eye mesh. It won't crash or
   anything, but the texture will be applied in an ugly-looking way - a
   typical UV mismatch.
The most common result of UV mismatches are seams - places where the
texture suddenly stretches and 'jumps' from one part to another. But
they can cause all kinds of weird-looking output, including what became
referred to as 'googly eyes'.

So that's the issue, as I understand it. What do we do to detect and fix
it? We check if a plugin changes the eye meshes used by a race without
also overriding the EYES record used by that race.

That doesn't actually suffice to fix it. For example, in the scenario
above, we would detect absolutely nothing wrong - because Mod A and
Mod B are still using the vanilla paths. We would assume everything was
fine, when in actuality googly eyes are awaiting us ingame.

Which brings us to the crux of this issue: *This is a resource
conflict!* You should solve this using BAIN, NifSkope and a DDS-capable
image viewer. Like any other resource conflict involving UV mismatches.
It has absolutely *no* business being in the BP.

Unless I'm completely misunderstanding it. But even if I am, this
patcher is horrid and breaks one of the most popular mods on the
Oblivion Nexus right now (OCO, to the point that an unofficial patch is
needed: https://www.nexusmods.com/oblivion/mods/49988).

And if I am misunderstanding it, then someone needs to provide a test
case. I asked on Discord and two of our biggest Oblivion experts could
not figure out what 'googly eyes' actually are, I had to go digging on
Google - so no guarantee that any of this is correct.

Perhaps the most enlightening thing they had to say was "I saw the Eye
Checker as only harmful and in no way helpful to me." When a 'feature'
is described like that by the experts in a field, it should probably go.

</rant>

Closes #655