Complain about inactive masters before merging
As well as about missing ones, but those should already be painfully
obvious due to the red checkbox.

The problem is that a merged plugin is usually inactive and so does not
force its masters to be active as well. However, if we want to merge it
into the BP, its masters have to be active, merged or missing (the
latter only for Filter plugins, of course). Inactive does not work and
will cause merging to crash.