Refactor Activate Plugins menu and lo_activate_all
Gets rid of ChoiceLink usage (MultiLink is much more appropriate here),
cleans up the _self = self weirdness, simplifies the API (menu is now
created directly by the Active Plugins link rather than by links_init),
etc.

This also changes the behavior of the Activate All link so that it
doesn't show an error popup but only an info popup if too many mergeable
plugins are present to activate all of them, since that's not a problem
the same way having too many non-mergeable plugins is. Also refactored
the method heavily and dropped the deprints, since those will display
error popups to the user when the error is actually properly handled by
us later on (barring the BoltError deprint, which I just moved up to the
link).

I then used the refactored method to implement a new link requested on
the Discord: Activate Non-Mergeable, which skips the non-mergeable
plugins entirely.