Rewrite lo_activate_exact to keep LO invariants
This method has always confused me because it's the only place in Wrye
Bash that can violate one of the key invariants of load order:

  Active(p) => ∀m ∈ masters(p): Active(m)

(i.e. if p is active, then all its masters must be active too)

The only other way to violate this invariant I can think of is to have
an active plugin with missing masters, but then you have an even bigger
problem (namely the game CTDing on the main menu) - and we rightfully
mark that situation with a giant red background! I can't fathom what
possible use case there is for activating a plugin without activating
its masters. It breaks the BP for one thing, since e.g. MobCell and
friends kind of rely on the masters having been loaded (see recent
tracebacks of the BP crashing with merged plugins that have inactive
masters - same thing happens here too, and fix is the same (i.e.
activating the darn masters)).

Also rename Activate Selected to Activate Only Selected to make it clear
that it doesn't just activate the selected plugins but *only* the
selected plugins.