Skip ENB image files on Screenshots tab
They're not screenshots and certainly shouldn't be managed by the
Screenshots tab. Took the opportunity to touch up the Screenshots tab's
part of the advanced docs - tweak the text to be game-agnostic, update
the example to highlight WB's "Next Shot" feature, add a table for
noting the difference between Morrowind's and all later games' INI
settings for screenshots and add docs for the new skips.