Goodbye, _shellUI
It has always confused me why we have two different looks and feels for
deletion which change depending on what tab you're using. Especially
because the _shellUI one is much less powerful (not showing a preview
or letting you uncheck items). And now that you can also adjust the
recycling state via the dialog, I don't know why we'd want to restrict
it to just the Mods tab. Plus takes some more lines of code along for
the ride in the already pretty complex deletion logic.