Account for CELL record header
This is probably not the right place to put it, but that was the issue:
the CELL record header was not getting included in the size.

BP is nearly identical now, with one exception: the records count in
the header and description is one higher than dev.

getNumRecords

m is an int (6), so it was thinking every single CELL had children. You
want the values, not the keys :P

getNumRecords pt2

Cell children groups weren't being counted correctly.

Interior cell BP is now perfect AFAICT.

Ut: fix WRLD children getNumRecords

Well I didn't copy and I updated with the deleted dict - hadn't used
_process_rec cause I would need to handle return values in there

Fix persistent CELL size calculation

Otherwise it'll be counted as a normal stray record, which will skip one
of the headers in the calculation, meaning we'll be off by exactly one
header size.

Ut: most of the issues here and some remaining subtle bugs will be
addressed in "Refactor getSize" commit - kept separate partly to
motivate that commit.