get_num_records:
getNumRecords is very hard to reason about so this is a first step
towards simplifying - note we only need a base implementation for this
one so the conceptual weight of the Mob API is not really increased -
that's good cause we do want to decrease it actually.
Mopy/bash/basher/mod_links.py: I used MobObjects.__bool__
Turns out there is only one use of get_num_records.