Rework pathcFile.getKeeper():
Use id_record in build patch - we might eventually want to use
iter_present_records there too (think also encapsulating id_records)
but for now let us catch the errors.

Refactor patch.keep so it also calls setChanged on records to keep, so
we will eventually get rid of the set_records_changed hack. We also
check here if the record should have been skipped, so we detect
possible bugs in previous phases of the patch.

This might finally put #653 to rest - with a deprint. At least the BP
should be safe as we intercept keep so no deleted in the patch.

Infernio: keep fixups

Found via "keep\([^,)]+\)" regex search.