Drop set_records_changed:
I did track load - older code (as in parsers) meticulously calls
setChanged and in the patch we should now be calling setChanged on
keeping the record. set_records_changed was not the way this API was
meant to be used (the way it's used in parsers is probably it) and adds
to the _AMobBase API so it should go, but the setChanged API is a bit
oldish style anyway - for instance, we have to make sure we call it.
Couple notes:

- changed and data should not be present `data is None == changed`
- but then as discussed in MreRecord.__init__ we might want to toss
data if we unpacked it anyways.

These are subtle matters to be addressed when we address loading of
plugins as part of #312.

Infernio: fix newly created records only writing headers

Mopy/bash/patcher/patch_files.py:
Since we weren't calling setChanged on them, they thought they had size
0 and so only dumped out a record header.