Drop handling of DarkPCB record:
If re-added should be an override of the _load_rec_group for the
respective type in _import_records (that deletes the record). For
discussion see:

https://github.com/wrye-bash/wrye-bash/commit/f4885f41b3e6ce81605692642628504ada742bd7

from where (thanks @XJDHDR)

[...]I checked this version plus the 3 stable patches Bethesda released:
- v1.1.425
- v1.1.511
- v1.2.214
- v1.2.416

There are no records with the FormIDs 0000A31D or 0004DECC in v1.1.425.
In v1.1.511, there is still no 0004DECC record, but Bethesda did add a
record which used 0000A31D: The script DarkPCB.

This remained the case for the two v1.2 patches as well. In other
words, record 0000A31D was still the script DarkPCB in v1.2.416 (the
latest version).

I then copied Oblivion.esm from my GOG copy of the game and opened it
in TES4Edit. In this case, I observed the same things as Infernio;
0000A31D was the FormID for a CELL in the SEWorld WRLD record, and
DarkPCB's FormID was now 0004DECC.

So for some reason, Bethesda decided to reassign the script's FormID
when they created SI. People running older versions is not the problem.
The problem is that this FormID has a fundamentally different meaning
depending on whether the user has SI installed or not.