WIP refactor merge_records:
loadSet was only used when doFilter was True - this is simpler but still
not simple

Note peculiar TopComplexGrup override (note also that the order of
operations is inverted?)

any FormIDs with missing masters -> any FormIDs with *inactive* masters?

Get rid of MasterSet local import:

Aka _that_ local import - and there was a reason it was not easy to
uproot - essentially the part of master set that prevented it from
landing in bolt is the load order dependency.
EDIT: actually getOrdered had a single use that I swapped for load_order
get_ordered permitting me to land this on bolt.

Note I moved the `if load_set` check outside of
_masters_miss_after_filtering to avoid the function call for most
plugins like the previous code was doing - maybe faster to check
`if load_set is not None` here -> yep:

3.11.1 (tags/v3.11.1:a7a450f, Dec  6 2022, 19:58:39) [MSC v.1934 64 bit (AMD64)]
if not none: pass: 0.0013265999732539058
if none is not None: pass: 0.0011183000169694424

loadSet -> load_dict:

And drop single use attribute loadMods

WIP naming