WIP centralize bashed patch mod arrays:
These are essentially very fast calculations that can land in init - so
by creating the BP a bit earlier we centralize the active mods handling
which was extremely difficult to follow between the BP, the dialog, and
the GUI - poc, for one this must be injected to the GUI, see next:

Use the BP in _PatcherPanel:

Mopy/bash/basher/gui_patchers.py: note we were checking with
present_plugins = set(bosh.modInfos) - now we only look at all_plugins
Apart from this difference the rest should be pure refactoring - note we
compute now most stuff (like 'Filter' in tags) *once*, executing_patch
gone and less calls to getBashTags - seems like the right thing.

MergePatchesPatcher - should we not filter also sources for inactive_inm
in MergerPanel(_ListPatcherPanel)?

Kept _bp private for it to stand out - we should pass it as an init
parameter to gui patchers, save we instantiate those once...