Drop data from _HeadedGrup:
As seen the _HeadedGrup expects iter_records to be able to iterate over
records so the idea of loading (children) grups as blobs won't work. If
we want to do that we must rewrite iter_records to actually load
MreRecord after the fact - too much of a hassle for now. This cleans
the last warnings on id_records and co - note _write_header remains in
_HeadedGrup.

I finally renamed data -> grup_blob - only record.data remains
essentially ;)