Fix handling of ex/interior_temp_extra:
Those apparently are merged by record signature - in updateRecords we
also copy the record if a record with the same signature exists:

updateRecords: Only update if it's already present in the patch.

merge_records: Update even if it's not present in the patch yet.

I wonder how would those (like PGRD or ROAD) end up in the patch however
in the first place? Added by scanLoadMods :lightbulb: ?