Centralize plugins loading:
Initially for initData - aka retire ModLoader, now the PatchFile is
responsible for the loading of masters and srcs (:music:). scanLoadMods
needs to load those for all/most/? plugins so if we want to share the
_loaded_mods cache we should add more mods to the factories. I actually
realized this while writing this commit, which is one of the goals,
making the BP process more transparent. The main gain here is that the
client code does not call load_plugin anymore just accesses a cache -
note LoadFactory/ModFile imports eliminated. And of course performance:

dev at ad85967c3ad81ad7d80d590a4124f8ea0a7b06fe
patcher_dialog.py  194 PatchExecute:
841102318 function calls (803321382 primitive calls) in 511.523 seconds
      845    0.053    0.000  349.348    0.413 bash\mod_files.py:157(load_plugin)

this
patcher_dialog.py  194 PatchExecute:
767411552 function calls (730724377 primitive calls) in 460.678 seconds
      641    0.036    0.000  304.380    0.475 bash\mod_files.py:157(load_plugin)

timings fluctuate (and have also the profiler overhead).

Note however that memory consumption went up - we need to lighten up
the records, which is one of the goals of removing setDefault. Note also
that I had to request GLOB loaded for the TimescaleCheckerPatcher, which
otherwise, if it needed to load the master esm, it would do so for
all the signatures - a nogo. We need to allow load_plugin to add to
loaded tops for a more flexible API.

Mopy/bash/patcher/base.py: re-add progress - or drop it as those csvs
are tiny - we could however move csvs handling right on the ListPatcher
(edit: went for dropping to simplify centralizing csv handling).

Mopy/bash/patcher/patchers/mergers.py:
- I dropped the `if srcMod not in minfs` checks as this is done in
_process_sources
- the cachedMasters mantra is now moved to the patch file
- note sometimes we iterate in reversed order some not (?)

Removed the APatcher._update_patcher_factories - not needed but if this
covers also scanModFile we may add it back

Under #312.