Show busy cursor when creating BP dialog
We can't speed this up at all - about 95% of the time used for creating
the BP dialog is spent constructing native wx widgets in C++. But we can
at least show a busy cursor to make it clear that we're doing work.

Busy cursor instead of a progress dialog because it usually takes ~0.2s
for me, much too fast for a progress dialog.