save_headers: Rework writing, reenable TES5 & FO4
Calculate the master map properly and earlier, then proceed with the
same idea as the cosaves (remap first, then write the new masters). Also
avoids trying to reassign SaveFileHeader.masters, which doesn't work for
saves with ESLs. And finally, fixes us writing out the length of unicode
strings by encoding first, then writing, then decoding again - wastes a
bit of performance, but writing saves is not an operation where
performance matters much (compared to reading, which we do on boot).