Support TOML files on the INI Edits tab
This format is (mostly) a standardized version of INIs. It does contain
some complicated syntax that makes it more like a hybrid between YAML
and INI, but we can pretty much ignore that for now since all the mods
that have started using it use it like a bog standard INI file.

Also contains support for inline comments in INI files, since TOML needs
that.

Closes #661