[!] Sort and tidy imports, drop cruft, decouple
Note that we originally added isort to the CI as well, but that just
turned into a bunch of noise. We can run it from time to time, but
trying to satisfy it on every commit seems overkill.

[!] because of all the moving of imports and breaking of circular
dependencies.

We require pywin32 for much more than just wizards, plus I don't even
think wizards require pywin32 (anymore?). Breaks quite a bit of
coupling in basher, which is nice.

Also drop the startupinfo assignment, we don't even use it outside of
bolt and omods.

Break coupling in bosh, move imports to top level

Makes it easier to see at a glance what imports what and allows easier
spotting of circular dependencies (marked a couple with '##:').

Get rid of a couple local imports

All the BAIN imports in bosh can go to the single top-level import in
the middle of the file - we're already hiding the circular dependency
there, might as well centralize that hack.

Also moved coveragerc into new pyproject.toml.

Drop tomli license from LICENSE-THIRD-PARTY

We now use the stdlib version, so no more need for this.