Fix BAIN wizards failing on compareSEVersion
Broke way back in 851a6a01db744a8fd26fbf90490c6720e44ce66d, what a
throwback. Watch me fail to understand that this returns a list with two
values, and that only the first one matters!

Also took this opportunity to clean up version handling in belt, getting
rid of all the float and str comparisons of versions in favor of the proper
solution: bolt.LooseVersion.

Mopy/bash/_games_lo.py:
Fix WB not fixing LOs with too many actives

Not sure if this was always broken, but it was easy to reproduce: Create
>255 dummy plugins, open plugins.txt in a text editor and enable the
dummy plugins in there. Then tab into WB and watch it get stuck in an
infinite loop of showing warning popups about correcting the incorrect
LO.

Mopy/Docs/Wrye Bash Version History.html:
Add the 32-bit/Win7 notes to the version history

We already have these on the GitHub releases page and on the Nexus
pages, so we should list them in the version history for ease of access
too.