Store auto-column-size settings per tab
Instead of storing this setting globally, include the keyPrefix when
storing it. Turns this from a landmine that will destroy your carefully
resized columns on an unrelated tab/view/etc. into a very useful
feature.

Big open question is Ctrl+Numpad Plus - that works and we've documented
it in the docs, but I have no clue *why* it works. Probably done by
wxWidgets or Windows internally for some reason. I did go ahead and fix
that shortcut not updating our auto-size setting though.

Closes #625
Under #190 for untangling the wxFlag mess