converters: Improve warning when applying fails
The usual reason this happens is because we've got two packages with the
same contents CRC, and the most common way (honestly the only way) that
happens is because a BCF only moved or renamed files, leaving the
cumulative CRC of its contents unchanged. We should mention that, and
also show details about why we failed to apply the BCF.

And since we now show those errors to the user, make them nicer to read
and translatable.

Mopy/bash/brec/basic_elements.py:
Fix minor docstring typo