AskDialogue(DialogWindow):
Well high time - this is the beginning of the end of #190 (along with
some work on a thing that's called ListBoxes). We 're no more than 100
commits away (edit: 200). Drop vistaDialog import from balt is more than
welcome.

WIP will be further worked in this branch - Is this ask_uac_restart
still needed?