Clean up header flags handling
Implement all missed FO4 header flags (so far) and drop all the mixins.
They make it much harder for me to see if I've implemented the header
flags correctly (and as a POC, doing this already caught two errors in
the header flags definitions). The line increase is neglibible (~40
lines right now).