Check for NULL fids and unnecessary deletions
Enhance the Plugin Checker to check for records beside the main header
having a FormID of 00000000, as requested on Discord. Also requested on
Discord was an enhancement for the Plugin Checker to make it detect new
records with the Deleted flag and not advise cleaning them, since xEdit
will skip them when cleaning. Instead, we now show those in a separate
category. Also took the opportunity to clean up and refactor a bit in
here.

Ignore mod index for NULL detection

Note that 01000000 is *not* a valid FormID in FO4, i.e. the valid
FormIDs with a version 1.0 plugin in FO4 begin at 0x001, not at 0x000.