Fix deactivating game master in timestamp games
Not sure *when* this broke, but it broke because _order_fixed was used
for both lord and acti and didn't care whether it ended up adding new
actives in the process of fixing plugins in place. Changing that to only
fix present plugins in place is fine, because _fix_load_order will have
already assigned LO to plugins that were missing it when we're dealing
with lord rather than acti.

Also fix wrong deprints when we reorder active plugins (before_reorder
took a reference to acti rather than acti[:], i.e. a copy of the list).