Centralize ListPatcher sources validation:
This was motivated by the CsvListPatcher._update_patcher_factories. I
could not be sure that the srcs attribute was not set to some list of
mods by mistake - so I had finally to add validation to the patcher.
Of course the logic was there but as seen in various bits and places,
most notably in gui_patchers - splitting responsibilities between the
APatcher and PatcherPanel was a long journey that came home finally -
PatcherPanel has now the responsibilities it should have, namely config
(as in pickled config) handling and GUI panels creation, and delegates
to the patcher for the business logic as it should. The patcher guards
its init method (via _process_sources) from being fed junk, which is a
step towards making the patcher headless (see #600) and therefore
testable. Note:

- autoKey (now patcher_tags) really belongs to ListPatcher. Certainly
did not belong to the gui patchers, hence some ugly get_cls_vars
dropped or simplified
- list_patches_dir centralized and the ugly global became a beautiful
instance attribute
- _csv_key belongs also to the ListPatcher but really to the parser -
what stands in the way here is CsvListPatcher
- all_tags cache not really needed (few calls of getBashTags really
it' more for clarity)
- some bush game vars uses reduced - another sign autoKey belonged to
the model
- we could change the logic of the validation to print some warnings
(for instance for mods tagged with add *and* remove spells) - and the
validation itself could use some more items but the tough part is done,
rest is nit
- game_specific_import_patchers did not seem to have any csv support
ever - keep it that way. Note _AListPanelCsv only remaining purpose
is listLabel - get rid of that class (done)
- in

@@ -474,5 +473,4 @@ def _getConfig(self, configs):
         #--Verify file existence
-        self.configItems = [srcPath for srcPath in self.configItems if (
-                srcPath in  or (srcPath.fn_ext == '.csv' and
-                                             srcPath in patches_set()))]
+        self.configItems = self.patcher_type.get_sources(self._bp,
+                                                         self.configItems)

we were checking for bosh.modInfos membership - this seems wrong. Anyway
decoupling gui_patchers from bosh is a laudable goal.
- _ListsMergerPanel._get_auto_items escapes me a bit
- actually the checks in MergePatchesPatcher._validate_mod are
essentially encapsulated in PatchFile.set_active_arrays (bp_mergeable)
as seen - well.

Mopy/bash/patcher/patch_files.py: I don't see why making the dict to a
set - rework the arrays a bit.