Handle other web errors when doing update check
Treat them the same way we treat status code errors. The most common way
for this to occur is if the user just doesn't have an internet
connection. Previously we would pop up a bunch of error windows on boot,
which is definitely not of any use to anyone. Instead, just ignore this
and hope the internet connection comes online later, then try again.

Regression fix for skip_local change

1d6db12224ed42fba92d2fdaee29142c16328b6d incorrectly inverted the
skip_local check. Whoops, only noticed once it was committed already.