Add support for conditions to update check
This way we can add download options that are only supported by certain
operating systems etc. For example, we could add a Chocolatey link in
the future and only show it on Windows.

Already used to hide the Nexus option on non-Windows platforms - on
those, you'll want to head to GitHub and grab the source code.

Technically adds packaging as an optional requirement, but packaging is
already included in our builds because everything under the sun requires
it.